+++ B0.ml
@@ -65,12 +65,15 @@ let update_completion_scripts =
   let zsh = B0_env.in_scope_dir env ~/"src/tool/zsh-completion.sh" in
   let ml = B0_env.in_scope_dir env ~/"src/tool/cmdliner_data.ml" in
   let* bash = Os.File.read bash in
+  let bash = String.replace_first ~sub:"_cmdliner_generic" ~by:"%s" bash in
   let* zsh = Os.File.read zsh in
+  let zsh = String.replace_first ~sub:"_cmdliner_generic" ~by:"%s" zsh in
   let src = Fmt.str
-      "let bash_generic_completion =\n{|%s\
-       |}\n\n\
-       let zsh_generic_completion =\n{|%s\
-       |}" bash zsh
+      "let strf = Printf.sprintf\n\n\
+       let bash_generic_completion fun_name = strf\n{|%s\
+       |} fun_name\n\n\
+       let zsh_generic_completion fun_name = strf\n{|%s\
+       |} fun_name" bash zsh
   in
   Os.File.write ~force:true ~make_path:false ml src